Refactor file/folder hierarchy
Grouping sub/components by page.
