+++ client/src/components/ActiveGame/PriceAlertPreview.css
