+++ client/src/components/ActiveGame/SubmitEmailForm.js
