+++ client/src/components/Home/AutoSuggestions.js
