+++ client/src/components/Unsubscribe/Unsubscribe.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { browserHistory } from 'react-router'
-import GamePreview from './GamePreview';
+import GamePreview from '../helper/GamePreview';
 import { Alert, Button, Modal, ModalHeader, ModalBody, ModalFooter } from 'reactstrap';
 import './Unsubscribe.css';
