--- client/src/components/UserSignUp.css
@@ -1,13 +1,0 @@
-#gameInput:focus {
-    width: 60%;
-    background: #faffbd;
-    color: black;
-    transition: 0.3s;
-}
-
-#emailInput:focus {
-    width: 60%;
-    background: #faffbd;
-    color: black;
-    transition: 0.3s;
-}
