+++ client/src/components/helper/GamePreview.js
